services/log-consumer/requirements.txt
# Kafka
kafka-python-ng==2.2.2

# MongoDB
pymongo==4.6.1
motor==3.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# For development
pytest==7.4.3
pytest-asyncio==0.21.1
